Allow to chain expressions
